refactor(game): extract GameLaunchService out of game_service facade
Final F slice. Move the launch-execution cluster out of the GameService
static facade into a new GameLaunchService collaborator
(services/game/game_launch_service.dart): launchGame, the platform launch
methods (_launchGameAndroid/_launchGameDesktop/_launchGameDesktopFromConfig
/_launchStandaloneAndroid/_launchStandaloneDesktop), emulator/core and
RetroArch-variant resolution helpers, the argument parsers, the
process-liveness checks (_isProcessRunning/_isDefaultEmulatorRunning/
_isProcessRunningUnix + public isEmulatorRunning/isProcessRunning), and
handleAppResumed. GameLaunchResult moves too and is re-exported from
game_service.dart so importers resolve it unchanged.

GameService keeps thin delegators for the frozen public API (launchGame,
isEmulatorRunning, isProcessRunning, handleAppResumed) and now holds only
the Android game-lifecycle listener + the onScreenStateChanged slot.
Moved bodies are byte-identical vs main modulo provably-equivalent
delegator-unwrap substitutions (_registerGameLaunch ->
GameSessionManager.registerGameLaunch, recordGamePlayed ->
FavoritesService.recordGamePlayed, endGameSession ->
GameSessionManager.endGameSession, launchedEmulatorExe ->
GameSessionManager.launchedEmulatorExe).

Dropped 9 now-unused facade imports + the dead _log field. Host
1339 -> 193 (a thin facade shell); new file ~1195. No public-API change,
no behavior change. Closes out F.
