+++ lib/services/game/game_launch_service.dart
@@ -1,0 +1,1195 @@
+import 'dart:io';
+import 'package:flutter_localization/flutter_localization.dart';
+import 'package:neostation/l10n/app_locale.dart';
+import 'package:path/path.dart' as path;
+import 'package:flutter/services.dart';
+import 'package:flutter/material.dart';
+import 'package:neostation/services/logger_service.dart';
+import '../../models/game_model.dart';
+import '../../models/system_model.dart';
+import '../../models/emulator_model.dart';
+import '../../models/core_emulator_model.dart';
+import '../../repositories/emulator_repository.dart';
+import '../config_service.dart';
+import '../android_service.dart';
+import '../launcher_service.dart';
+import 'favorites_service.dart';
+import 'game_session_manager.dart';
+import '../gamepad/gamepad_navigation_manager.dart';
+
+/// Represents the result of a game launch attempt.
+/// Represents the result of a game launch attempt.
+class GameLaunchResult {
+  /// Whether the launch was successful.
+  final bool success;
+
+  /// Human-readable error message.
+  final String? errorMessage;
+
+  /// Technical details or raw error information for debugging.
+  final String? errorDetails;
+
+  GameLaunchResult.success()
+    : success = true,
+      errorMessage = null,
+      errorDetails = null;
+
+  GameLaunchResult.failure(this.errorMessage, [this.errorDetails])
+    : success = false;
+}
+
+/// Executes game launches across all supported platforms and monitors the
+/// launched emulator/game process.
+///
+/// Owns the platform-specific launch logic (RetroArch cores + standalone
+/// emulators, Android + desktop), emulator/core resolution, RetroArch variant
+/// resolution, and the process-liveness checks. Extracted verbatim from
+/// [GameService], which now delegates its launch API here. Registers each
+/// session via [GameSessionManager] and records plays via [FavoritesService].
+class GameLaunchService {
+  GameLaunchService._();
+
+  static final _log = LoggerService.instance;
+
+  /// Core logic for launching a game session across all supported platforms.
+  ///
+  /// Performs pre-launch validations (ROM existence, system config), resolves the
+  /// optimal emulator/player, and initiates the execution process.
+  static Future<GameLaunchResult> launchGame(
+    BuildContext context,
+    SystemModel system,
+    GameModel game,
+  ) async {
+    try {
+      if (Platform.isAndroid && (system.folderName == 'android')) {
+        if (game.romPath == null) {
+          return GameLaunchResult.failure(
+            AppLocale.packageNameMissing.getString(context),
+          );
+        }
+
+        GameSessionManager.registerGameLaunch(system, game, 'android_app');
+        await FavoritesService.recordGamePlayed(game);
+
+        final success = await AndroidService.launchPackage(game.romPath!);
+        if (!context.mounted) return GameLaunchResult.failure('', '');
+        if (success) {
+          return GameLaunchResult.success();
+        } else {
+          return GameLaunchResult.failure(
+            AppLocale.failedToLaunchAndroidApp.getString(context),
+            game.romPath,
+          );
+        }
+      }
+
+      bool romExists = false;
+      if (game.romPath != null) {
+        if (Platform.isAndroid && game.romPath!.startsWith('content://')) {
+          romExists = true;
+        } else {
+          romExists = await File(game.romPath!).exists();
+        }
+      }
+      if (!context.mounted) return GameLaunchResult.failure('', '');
+
+      if (!romExists) {
+        return GameLaunchResult.failure(
+          AppLocale.romFileNotFound.getString(context),
+          game.romPath ?? AppLocale.noData.getString(context),
+        );
+      }
+
+      final configFileName = '${system.folderName}.json';
+      final bool configLoaded = await LauncherService.instance.loadSystemConfig(
+        configFileName,
+      );
+
+      if (configLoaded) {
+        String? preferredPlayerId = game.emulatorName;
+
+        if (preferredPlayerId == null) {
+          final defaultEmu =
+              await EmulatorRepository.getDefaultEmulatorForSystem(system.id!);
+          if (defaultEmu != null) {
+            preferredPlayerId = defaultEmu.uniqueId;
+          }
+        }
+
+        final launchCmd = LauncherService.instance.getLaunchCommand(
+          system,
+          game,
+          preferredPlayerId,
+        );
+
+        if (launchCmd.isNotEmpty) {
+          if (Platform.isAndroid &&
+              launchCmd.containsKey('package') &&
+              launchCmd.containsKey('activity')) {
+            const platform = MethodChannel('com.neogamelab.neostation/game');
+
+            // Cleanup + real failure surfaced to the user. A failure in this
+            // branch means the user's *chosen* emulator (resolved from its JSON
+            // config) was actually attempted and could not launch — surface it
+            // instead of falling through to the generic standalone fallback
+            // below, which would launch a *different* emulator and mask the
+            // misconfiguration.
+            Future<GameLaunchResult> failLaunch() async {
+              GamepadNavigationManager.reactivate();
+              await platform.invokeMethod('setGamepadBlock', {'block': false});
+              if (!context.mounted) return GameLaunchResult.failure('', '');
+              return GameLaunchResult.failure(
+                AppLocale.launchFailed.getString(context),
+                AppLocale.error.getString(context),
+              );
+            }
+
+            try {
+              GamepadNavigationManager.reactivate();
+              await platform.invokeMethod('setGamepadBlock', {'block': true});
+
+              final activityName = launchCmd['activity'];
+              final action = launchCmd['action'];
+              final category = launchCmd['category'];
+              final data = launchCmd['data'];
+              final type = launchCmd['type'];
+
+              final List<Map<String, dynamic>> extrasList = [];
+
+              if (launchCmd.containsKey('extras') &&
+                  launchCmd['extras'] is List) {
+                for (final item in launchCmd['extras'] as List) {
+                  if (item is Map) {
+                    extrasList.add(Map<String, dynamic>.from(item));
+                  }
+                }
+              } else {
+                final argsStr = launchCmd['args']?.toString() ?? '';
+                if (argsStr.isNotEmpty) {
+                  final extrasMap = _parseArgsToExtras(argsStr);
+                  extrasMap.forEach((k, v) {
+                    String type = 'string';
+                    if (v is int) type = 'int';
+                    if (v is bool) type = 'bool';
+                    extrasList.add({'key': k, 'value': v, 'type': type});
+                  });
+                }
+              }
+
+              final packageName = await _resolveRetroArchVariant(
+                launchCmd['package'].toString(),
+                extrasList,
+                system,
+              );
+
+              final result = await platform
+                  .invokeMethod('launchGenericIntent', {
+                    'package': packageName,
+                    'activity': activityName,
+                    'action': action,
+                    'category': category,
+                    'data': data,
+                    'type': type,
+                    'extras': extrasList,
+                    'activity_flags': launchCmd['activity_flags'] != null
+                        ? List<String>.from(launchCmd['activity_flags'] as List)
+                        : <String>[],
+                    'keep_saf_uri': launchCmd['keep_saf_uri'] == true,
+                  });
+
+              if (result == true) {
+                GameSessionManager.registerGameLaunch(system, game);
+                await FavoritesService.recordGamePlayed(game);
+                return GameLaunchResult.success();
+              }
+              return await failLaunch();
+            } catch (e) {
+              _log.e('JSON Launch Error: $e');
+              return await failLaunch();
+            }
+          }
+
+          if ((Platform.isWindows || Platform.isLinux || Platform.isMacOS) &&
+              launchCmd.containsKey('executable')) {
+            if (!context.mounted) return GameLaunchResult.failure('', '');
+            return await _launchGameDesktopFromConfig(
+              context,
+              launchCmd,
+              system,
+              game,
+            );
+          }
+        }
+      }
+
+      final standaloneEmulator = await _getStandaloneEmulatorForSystem(system);
+      if (!context.mounted) return GameLaunchResult.failure('', '');
+
+      if (standaloneEmulator != null) {
+        if (Platform.isAndroid) {
+          return await _launchStandaloneAndroid(
+            context,
+            system,
+            game,
+            standaloneEmulator,
+          );
+        } else if (Platform.isWindows || Platform.isLinux || Platform.isMacOS) {
+          return await _launchStandaloneDesktop(
+            context,
+            system,
+            game,
+            standaloneEmulator,
+          );
+        } else {
+          return GameLaunchResult.failure(
+            AppLocale.platformNotSupported.getString(context),
+            Platform.operatingSystem,
+          );
+        }
+      }
+
+      final coreName = await _getCoreForSystem(system);
+      if (!context.mounted) return GameLaunchResult.failure('', '');
+
+      if (coreName == null) {
+        return GameLaunchResult.failure(
+          AppLocale.coreNotConfigured.getString(context),
+          'No core found for system ${system.folderName}',
+        );
+      }
+
+      if (Platform.isAndroid) {
+        return await _launchGameAndroid(context, system, game, coreName);
+      } else if (Platform.isWindows || Platform.isLinux || Platform.isMacOS) {
+        return await _launchGameDesktop(context, system, game, coreName);
+      } else {
+        _log.e('Platform not supported: ${Platform.operatingSystem}');
+        return GameLaunchResult.failure(
+          AppLocale.platformNotSupported.getString(context),
+          Platform.operatingSystem,
+        );
+      }
+    } catch (e) {
+      _log.e('Error launching the game: $e');
+      GamepadNavigationManager.reactivate();
+      if (Platform.isAndroid) {
+        const platform = MethodChannel('com.neogamelab.neostation/game');
+        await platform.invokeMethod('setGamepadBlock', {'block': false});
+      }
+      if (!context.mounted) return GameLaunchResult.failure('', '');
+      return GameLaunchResult.failure(
+        AppLocale.anErrorOccurred.getString(context),
+        e.toString(),
+      );
+    }
+  }
+
+  /// Internal logic for launching RetroArch cores on Android.
+  static Future<GameLaunchResult> _launchGameAndroid(
+    BuildContext context,
+    SystemModel system,
+    GameModel game,
+    String coreName,
+  ) async {
+    try {
+      final packages = await EmulatorRepository.getAndroidRetroArchPackages();
+
+      if (packages.isNotEmpty) {
+        try {
+          final defaultEmu =
+              await EmulatorRepository.getDefaultEmulatorForSystem(system.id!);
+          if (defaultEmu != null &&
+              defaultEmu.androidPackageName != null &&
+              defaultEmu.androidPackageName!.isNotEmpty) {
+            final specificPackage = defaultEmu.androidPackageName!;
+            _log.i(
+              'Android: User selected specific RetroArch package: $specificPackage',
+            );
+            packages.insert(0, specificPackage);
+          }
+        } catch (e) {
+          _log.e('Error getting default emulator package: $e');
+        }
+      }
+
+      const platform = MethodChannel('com.neogamelab.neostation/game');
+      GamepadNavigationManager.reactivate();
+      await platform.invokeMethod('setGamepadBlock', {'block': true});
+
+      String packageName = 'com.retroarch.aarch64';
+      if (packages.isNotEmpty) {
+        packageName = packages.first;
+      }
+
+      final activityName =
+          'com.retroarch.browser.retroactivity.RetroActivityFuture';
+
+      final result = await platform.invokeMethod('launchGenericIntent', {
+        'package': packageName,
+        'activity': activityName,
+        'action': 'android.intent.action.MAIN',
+        'category': 'android.intent.category.LAUNCHER',
+        'extras': [
+          {
+            'key': 'ROM',
+            'value': (game.romPath?.startsWith('content://') == true)
+                ? 'neostation-realpath:${game.romPath}'
+                : game.romPath ?? '',
+            'type': 'string',
+          },
+          {'key': 'LIBRETRO', 'value': coreName, 'type': 'string'},
+        ],
+      });
+
+      if (result == true) {
+        await FavoritesService.recordGamePlayed(game);
+        GameSessionManager.registerGameLaunch(system, game);
+        return GameLaunchResult.success();
+      } else {
+        GamepadNavigationManager.reactivate();
+        const platform = MethodChannel('com.neogamelab.neostation/game');
+        await platform.invokeMethod('setGamepadBlock', {'block': false});
+        if (!context.mounted) return GameLaunchResult.failure('', '');
+        return GameLaunchResult.failure(
+          AppLocale.launchFailed.getString(context),
+          'RetroArch returned false',
+        );
+      }
+    } catch (e) {
+      GamepadNavigationManager.reactivate();
+      const platform = MethodChannel('com.neogamelab.neostation/game');
+      await platform.invokeMethod('setGamepadBlock', {'block': false});
+      if (!context.mounted) return GameLaunchResult.failure('', '');
+
+      if (e is PlatformException) {
+        if (e.code == "CORE_NOT_FOUND") {
+          return GameLaunchResult.failure(
+            AppLocale.coreNotInstalled
+                .getString(context)
+                .replaceFirst('{name}', coreName),
+            'Please install the core from RetroArch\'s Online Updater',
+          );
+        } else {
+          return GameLaunchResult.failure(
+            e.message ?? AppLocale.error.getString(context),
+            e.details?.toString(),
+          );
+        }
+      } else {
+        _log.e('Error en MethodChannel: $e');
+        return GameLaunchResult.failure(
+          AppLocale.launchFailed.getString(context),
+          e.toString(),
+        );
+      }
+    }
+  }
+
+  /// Internal logic for launching RetroArch cores on Desktop.
+  static Future<GameLaunchResult> _launchGameDesktop(
+    BuildContext context,
+    SystemModel system,
+    GameModel game,
+    String coreName,
+  ) async {
+    try {
+      final detectedEmulators =
+          await EmulatorRepository.getUserDetectedEmulators();
+      final retroArch = detectedEmulators['RetroArch'];
+      if (!context.mounted) return GameLaunchResult.failure('', '');
+      if (retroArch == null) {
+        return GameLaunchResult.failure(
+          AppLocale.retroArchNotFound.getString(context),
+          'RetroArch is not detected on your system. Please install RetroArch.',
+        );
+      }
+
+      bool raExists = false;
+      if (Platform.isMacOS && retroArch.path.endsWith('.app')) {
+        raExists = await Directory(retroArch.path).exists();
+      } else {
+        raExists = await File(retroArch.path).exists();
+      }
+      if (!context.mounted) return GameLaunchResult.failure('', '');
+
+      if (!raExists) {
+        _log.e(
+          'RetroArch does not exist at the specified path: ${retroArch.path}',
+        );
+        return GameLaunchResult.failure(
+          AppLocale.retroArchExecutableNotFound.getString(context),
+          'Path: ${retroArch.path}',
+        );
+      }
+
+      final coresPath = await _getRetroArchCoresDirectory(retroArch);
+      final coresDirectory = Directory(coresPath);
+      if (!await coresDirectory.exists()) {
+        if (!context.mounted) return GameLaunchResult.failure('', '');
+        _log.e('The cores directory does not exist: $coresPath');
+        return GameLaunchResult.failure(
+          AppLocale.coresDirectoryNotFound.getString(context),
+          'Path: $coresPath',
+        );
+      }
+
+      final coreFullPath = await _getCoreFullPath(coreName);
+      if (!context.mounted) return GameLaunchResult.failure('', '');
+      if (coreFullPath == null) {
+        return GameLaunchResult.failure(
+          AppLocale.coreNotFound.getString(context),
+          'Could not locate core: $coreName',
+        );
+      }
+
+      final coreFile = File(coreFullPath);
+      if (!await coreFile.exists()) {
+        if (!context.mounted) return GameLaunchResult.failure('', '');
+        return GameLaunchResult.failure(
+          AppLocale.coreFileNotFound.getString(context),
+          'Path: $coreFullPath',
+        );
+      }
+
+      Process process;
+
+      if (Platform.isMacOS) {
+        String executable = retroArch.path;
+        if (executable.endsWith('.app')) {
+          executable = path.join(executable, 'Contents', 'MacOS', 'RetroArch');
+        }
+
+        final args = ['-L', coreFullPath, game.romPath!];
+        final env = Map<String, String>.from(Platform.environment);
+        env['HOME'] = ConfigService.getRealHomePath();
+
+        process = await Process.start(executable, args, environment: env);
+      } else {
+        String executable = retroArch.path;
+        final args = ['-f', '-L', coreFullPath, game.romPath!];
+
+        process = await Process.start(executable, args);
+      }
+
+      process.stdout.listen((_) {});
+      process.stderr.listen((_) {});
+
+      GamepadNavigationManager.deactivateAll();
+
+      process.exitCode
+          .then((exitCode) async {
+            _log.i('RetroArch exited with code: $exitCode');
+            await Future.delayed(Duration(seconds: 2));
+            bool stillRunning = await _isDefaultEmulatorRunning();
+
+            if (!stillRunning &&
+                (Platform.isWindows || Platform.isLinux || Platform.isMacOS)) {
+              GameSessionManager.endGameSession();
+            }
+          })
+          .catchError((error) {
+            _log.e('Error monitoring RetroArch: $error');
+            if (Platform.isWindows || Platform.isLinux || Platform.isMacOS) {
+              GameSessionManager.endGameSession();
+            }
+          });
+
+      await FavoritesService.recordGamePlayed(game);
+      GameSessionManager.registerGameLaunch(system, game);
+
+      return GameLaunchResult.success();
+    } catch (e) {
+      _log.e('Error launching game on ${Platform.operatingSystem}: $e');
+      GamepadNavigationManager.reactivate();
+      if (!context.mounted) return GameLaunchResult.failure('', '');
+      return GameLaunchResult.failure(
+        AppLocale.failedToLaunchRetroArch.getString(context),
+        e.toString(),
+      );
+    }
+  }
+
+  /// Internal logic for launching games using a custom JSON configuration profile.
+  static Future<GameLaunchResult> _launchGameDesktopFromConfig(
+    BuildContext context,
+    Map<String, dynamic> launchCmd,
+    SystemModel system,
+    GameModel game,
+  ) async {
+    try {
+      String executable = launchCmd['executable'].toString();
+
+      if ((Platform.isWindows || Platform.isLinux)) {
+        final detected = await EmulatorRepository.getUserDetectedEmulators();
+
+        if (executable.toLowerCase().contains('retroarch')) {
+          final ra = detected['RetroArch'];
+          if (ra != null && ra.path.isNotEmpty) {
+            if (executable != ra.path) {
+              _log.i(
+                'Resolving RetroArch executable from "$executable" to user-configured path: ${ra.path}',
+              );
+            }
+            executable = ra.path;
+          }
+        } else if (!await File(executable).exists()) {
+          String? resolvedPath;
+          final uniqueId = launchCmd['unique_id']?.toString();
+          if (uniqueId != null) {
+            for (final emu in detected.values) {
+              if (emu.uniqueId == uniqueId &&
+                  emu.detected &&
+                  emu.path.isNotEmpty) {
+                resolvedPath = emu.path;
+                break;
+              }
+            }
+          }
+
+          if (resolvedPath == null) {
+            final playerName = launchCmd['player_name']?.toString();
+            if (playerName != null) {
+              final emu = detected[playerName];
+              if (emu != null && emu.detected && emu.path.isNotEmpty) {
+                resolvedPath = emu.path;
+              }
+            }
+          }
+
+          if (resolvedPath != null) {
+            executable = resolvedPath;
+          }
+        }
+      } else if (Platform.isMacOS &&
+          executable.toLowerCase().contains('retroarch')) {
+        final detected = await EmulatorRepository.getUserDetectedEmulators();
+        final ra = detected['RetroArch'];
+        if (ra != null && ra.path.isNotEmpty) {
+          if (executable != ra.path) {
+            _log.i(
+              'Resolving RetroArch executable on macOS from "$executable" to user-configured path: ${ra.path}',
+            );
+          }
+          executable = ra.path;
+          if (executable.endsWith('.app')) {
+            executable = path.join(
+              executable,
+              'Contents',
+              'MacOS',
+              'RetroArch',
+            );
+          }
+        }
+      }
+
+      if ((Platform.isWindows || Platform.isLinux || Platform.isMacOS) &&
+          !await File(executable).exists()) {
+        if (!context.mounted) return GameLaunchResult.failure('', '');
+        _log.e('Final check failed: $executable not found');
+        return GameLaunchResult.failure(
+          AppLocale.executableNotFound.getString(context),
+          'Could not find the emulator or game executable at:\n$executable\n\nPlease check your System Settings or Emulator Configuration.',
+        );
+      }
+
+      final argsStr = launchCmd['args']?.toString() ?? '';
+      final args = LauncherService.splitArgs(argsStr);
+
+      final env = Map<String, String>.from(Platform.environment);
+      if (Platform.isMacOS) {
+        env['HOME'] = ConfigService.getRealHomePath();
+      }
+
+      final process = await Process.start(executable, args, environment: env);
+
+      process.stdout.listen((_) {});
+      process.stderr.listen((_) {});
+
+      GamepadNavigationManager.deactivateAll();
+
+      process.exitCode
+          .then((exitCode) async {
+            _log.i('Process exited with code: $exitCode');
+            await Future.delayed(Duration(seconds: 2));
+            bool stillRunning = false;
+            if (GameSessionManager.launchedEmulatorExe != null) {
+              stillRunning = await _isProcessRunning(
+                GameSessionManager.launchedEmulatorExe!,
+              );
+            }
+
+            if (!stillRunning &&
+                (Platform.isWindows || Platform.isLinux || Platform.isMacOS)) {
+              GameSessionManager.endGameSession();
+            }
+          })
+          .catchError((err) {
+            if (Platform.isWindows || Platform.isLinux || Platform.isMacOS) {
+              GameSessionManager.endGameSession();
+            }
+          });
+
+      await FavoritesService.recordGamePlayed(game);
+
+      String? launchedExeName;
+      if (!executable.toLowerCase().contains('retroarch')) {
+        launchedExeName = path.basename(executable);
+      }
+
+      GameSessionManager.registerGameLaunch(system, game, launchedExeName);
+
+      return GameLaunchResult.success();
+    } catch (e) {
+      _log.e('Error launching via config: $e');
+      GamepadNavigationManager.reactivate();
+      if (!context.mounted) return GameLaunchResult.failure('', '');
+      return GameLaunchResult.failure(
+        AppLocale.launchFailed.getString(context),
+        e.toString(),
+      );
+    }
+  }
+
+  /// Resolves the identifier for the core assigned to the given system.
+  static Future<String?> _getCoreForSystem(SystemModel system) async {
+    final emulator = await EmulatorRepository.getDefaultEmulatorForSystem(
+      system.id!,
+    );
+
+    if (emulator != null) {
+      final coreFilename = emulator['core_filename'].toString();
+
+      if (Platform.isAndroid) {
+        return coreFilename;
+      } else {
+        String coreName = coreFilename;
+        if (coreName.endsWith('.dll')) {
+          coreName = coreName.substring(0, coreName.length - 4);
+        } else if (coreName.endsWith('.so')) {
+          coreName = coreName.substring(0, coreName.length - 3);
+        }
+
+        return coreName;
+      }
+    }
+
+    _log.e('No default emulator found for system ${system.folderName}');
+    return null;
+  }
+
+  /// Retrieves the user-assigned standalone emulator for a system if applicable.
+  static Future<Map<String, dynamic>?> _getStandaloneEmulatorForSystem(
+    SystemModel system,
+  ) async {
+    if (system.id == null) return null;
+
+    try {
+      final standalones =
+          await EmulatorRepository.getStandaloneEmulatorsBySystemId(system.id!);
+
+      if (standalones.isEmpty) {
+        return null;
+      }
+
+      Map<String, dynamic>? userDefault;
+      for (final standalone in standalones) {
+        if (standalone['is_user_default'] == 1) {
+          userDefault = standalone;
+          break;
+        }
+      }
+
+      if (userDefault == null) {
+        return null;
+      }
+
+      if (!Platform.isAndroid) {
+        final path = userDefault['emulator_path']?.toString();
+        if (path == null || path.isEmpty) {
+          return null;
+        }
+
+        final file = File(path);
+        if (!await file.exists()) {
+          return null;
+        }
+      }
+
+      return userDefault;
+    } catch (e) {
+      _log.e('Error getting standalone emulator: $e');
+      return null;
+    }
+  }
+
+  /// Internal logic for launching standalone emulators on Android.
+  static Future<GameLaunchResult> _launchStandaloneAndroid(
+    BuildContext context,
+    SystemModel system,
+    GameModel game,
+    Map<String, dynamic> emulator,
+  ) async {
+    try {
+      final packageName = emulator['android_package_name']?.toString();
+      final activityName = emulator['android_activity_name']?.toString();
+
+      if (packageName == null || packageName.isEmpty) {
+        _log.e('Missing Android package for standalone emulator');
+        return GameLaunchResult.failure(
+          AppLocale.emulatorNotConfigured.getString(context),
+          'Missing Android package name for ${emulator['name']}',
+        );
+      }
+
+      if (activityName == null || activityName.isEmpty) {
+        _log.w(
+          'Standalone emulator ${emulator['name']} has no activity; falling back to package launcher',
+        );
+      }
+
+      const platform = MethodChannel('com.neogamelab.neostation/game');
+      GamepadNavigationManager.reactivate();
+      await platform.invokeMethod('setGamepadBlock', {'block': true});
+
+      final romPath = game.romPath!;
+      final dataUri =
+          romPath.startsWith('content://') || romPath.startsWith('file://')
+          ? romPath
+          : Uri.file(romPath).toString();
+
+      final result = await platform.invokeMethod('launchGenericIntent', {
+        'package': packageName,
+        'activity': activityName,
+        'action': 'android.intent.action.VIEW',
+        'data': dataUri,
+        'activity_flags': <String>[],
+        'keep_saf_uri': false,
+      });
+
+      if (result == true) {
+        await FavoritesService.recordGamePlayed(game);
+        GameSessionManager.registerGameLaunch(system, game);
+        return GameLaunchResult.success();
+      } else {
+        _log.e('Failed to launch standalone emulator on Android');
+        GamepadNavigationManager.reactivate();
+        await platform.invokeMethod('setGamepadBlock', {'block': false});
+        if (!context.mounted) return GameLaunchResult.failure('', '');
+        return GameLaunchResult.failure(
+          AppLocale.failedToLaunchStandalone
+              .getString(context)
+              .replaceFirst('{name}', emulator['name']),
+          'The emulator may not be installed or the app is not responding',
+        );
+      }
+    } catch (e) {
+      _log.e('Error launching standalone emulator on Android: $e');
+      GamepadNavigationManager.reactivate();
+      const platform = MethodChannel('com.neogamelab.neostation/game');
+      await platform.invokeMethod('setGamepadBlock', {'block': false});
+      if (!context.mounted) return GameLaunchResult.failure('', '');
+
+      if (e is PlatformException) {
+        return GameLaunchResult.failure(
+          e.message ?? 'Platform error',
+          'Code: ${e.code}\nDetails: ${e.details}',
+        );
+      }
+      return GameLaunchResult.failure(
+        AppLocale.launchFailed.getString(context),
+        e.toString(),
+      );
+    }
+  }
+
+  /// Internal logic for launching standalone emulators on Desktop.
+  static Future<GameLaunchResult> _launchStandaloneDesktop(
+    BuildContext context,
+    SystemModel system,
+    GameModel game,
+    Map<String, dynamic> emulator,
+  ) async {
+    try {
+      final emulatorPath = emulator['emulator_path']?.toString();
+      final launchArgs =
+          emulator['launch_arguments']?.toString() ?? '"{rom_path}"';
+
+      if (emulatorPath == null || emulatorPath.isEmpty) {
+        _log.e('Emulator path not configured');
+        return GameLaunchResult.failure(
+          AppLocale.emulatorNotConfigured.getString(context),
+          'Path not set for ${emulator['name']}',
+        );
+      }
+
+      final emulatorFile = File(emulatorPath);
+      if (!await emulatorFile.exists()) {
+        if (!context.mounted) return GameLaunchResult.failure('', '');
+        _log.e('Emulator not found: $emulatorPath');
+        return GameLaunchResult.failure(
+          AppLocale.executableNotFound.getString(context),
+          'Path: $emulatorPath',
+        );
+      }
+
+      final romPath = game.romPath!;
+      final args = launchArgs
+          .replaceAll('{rom_path}', romPath)
+          .replaceAll('{emulator_path}', emulatorPath);
+
+      final argList = _parseCommandArguments(args);
+      final process = await Process.start(emulatorPath, argList);
+
+      process.stdout.listen((_) {});
+      process.stderr.listen((_) {});
+
+      GamepadNavigationManager.deactivateAll();
+
+      process.exitCode
+          .then((exitCode) async {
+            _log.i('Standalone emulator exited with code: $exitCode');
+            await Future.delayed(Duration(seconds: 2));
+            bool stillRunning = false;
+            if (GameSessionManager.launchedEmulatorExe != null) {
+              stillRunning = await _isProcessRunning(
+                GameSessionManager.launchedEmulatorExe!,
+              );
+            }
+
+            if (!stillRunning &&
+                (Platform.isWindows || Platform.isLinux || Platform.isMacOS)) {
+              GameSessionManager.endGameSession();
+            }
+          })
+          .catchError((error) {
+            _log.e('Error monitoring standalone emulator: $error');
+            if (Platform.isWindows || Platform.isLinux || Platform.isMacOS) {
+              GameSessionManager.endGameSession();
+            }
+          });
+
+      final exeName = path.basename(emulatorPath);
+      GameSessionManager.registerGameLaunch(system, game, exeName);
+      await FavoritesService.recordGamePlayed(game);
+
+      return GameLaunchResult.success();
+    } catch (e) {
+      _log.e('Error launching standalone emulator: $e');
+      GamepadNavigationManager.reactivate();
+      if (!context.mounted) return GameLaunchResult.failure('', '');
+      return GameLaunchResult.failure(
+        AppLocale.failedToLaunchStandalone
+            .getString(context)
+            .replaceFirst('{name}', emulator['name']),
+        e.toString(),
+      );
+    }
+  }
+
+  /// Parses a command-line argument string into an Android Intent extras map.
+  /// Resolves the concrete RetroArch package variant for a JSON launch.
+  ///
+  /// The system JSON hardcodes the base `com.retroarch` package, but the user
+  /// may have a different variant installed (e.g. `com.retroarch.aarch64`),
+  /// stored on their default emulator record. All RetroArch variants share the
+  /// same launch activity, so only the package and the `CONFIGFILE` extra's
+  /// path need patching — substituting a *standalone* emulator's package here
+  /// would instead produce an unresolvable intent (standalone package +
+  /// RetroArch activity → ActivityNotFound), so the substitution is gated on
+  /// the default emulator itself being a RetroArch variant.
+  ///
+  /// Returns the package to launch and mutates [extras] in place. For non-
+  /// RetroArch intents it returns [package] unchanged and leaves [extras] alone.
+  static Future<String> _resolveRetroArchVariant(
+    String package,
+    List<Map<String, dynamic>> extras,
+    SystemModel system,
+  ) async {
+    if (!package.startsWith(CoreEmulatorModel.retroArchPackagePrefix)) {
+      return package;
+    }
+
+    var resolved = package;
+    try {
+      final defaultEmu = await EmulatorRepository.getDefaultEmulatorForSystem(
+        system.id!,
+      );
+      if (defaultEmu != null && defaultEmu.isRetroArch) {
+        resolved = defaultEmu.androidPackageName!;
+      }
+    } catch (e) {
+      _log.e('Error resolving RetroArch package variant: $e');
+    }
+
+    // Point the RetroArch CONFIGFILE extra at the resolved variant's config
+    // directory (only the non-base variants live under their own package dir).
+    if (resolved != CoreEmulatorModel.retroArchPackagePrefix) {
+      for (final extra in extras) {
+        if (extra['key'] == 'CONFIGFILE') {
+          final currentPath = extra['value'].toString();
+          if (currentPath.contains('/com.retroarch/')) {
+            extra['value'] = currentPath.replaceAll(
+              '/com.retroarch/',
+              '/$resolved/',
+            );
+          }
+        }
+      }
+    }
+
+    return resolved;
+  }
+
+  static Map<String, dynamic> _parseArgsToExtras(String argsStr) {
+    if (argsStr.isEmpty) return {};
+
+    final extras = <String, dynamic>{};
+    final args = _parseCommandArguments(argsStr);
+
+    for (int i = 0; i < args.length; i++) {
+      final arg = args[i];
+      if (arg == '-e' || arg == '--es') {
+        if (i + 2 < args.length) {
+          extras[args[i + 1]] = args[i + 2];
+          i += 2;
+        }
+      } else if (arg == '--ez') {
+        if (i + 2 < args.length) {
+          extras[args[i + 1]] = args[i + 2] == 'true' || args[i + 2] == '1';
+          i += 2;
+        }
+      } else if (arg == '--ei') {
+        if (i + 2 < args.length) {
+          extras[args[i + 1]] = int.tryParse(args[i + 2]) ?? 0;
+          i += 2;
+        }
+      } else if (arg == '--esa') {
+        if (i + 2 < args.length) {
+          extras[args[i + 1]] = args[i + 2]
+              .split(',')
+              .map((e) => e.trim())
+              .toList();
+          i += 2;
+        }
+      }
+    }
+    return extras;
+  }
+
+  /// Tokenizes a command string into discrete arguments, respecting double quotes.
+  static List<String> _parseCommandArguments(String args) {
+    final List<String> result = [];
+    final buffer = StringBuffer();
+    bool inQuotes = false;
+
+    for (int i = 0; i < args.length; i++) {
+      final char = args[i];
+
+      if (char == '"') {
+        inQuotes = !inQuotes;
+      } else if (char == ' ' && !inQuotes) {
+        if (buffer.isNotEmpty) {
+          result.add(buffer.toString());
+          buffer.clear();
+        }
+      } else {
+        buffer.write(char);
+      }
+    }
+
+    if (buffer.isNotEmpty) {
+      result.add(buffer.toString());
+    }
+
+    return result;
+  }
+
+  /// Resolves the absolute path for a specific RetroArch core library.
+  static Future<String?> _getCoreFullPath(String coreName) async {
+    try {
+      final detectedEmulators =
+          await EmulatorRepository.getUserDetectedEmulators();
+      final retroArch = detectedEmulators['RetroArch'];
+      if (retroArch == null) return null;
+
+      final coresDir = await _getRetroArchCoresDirectory(retroArch);
+
+      String fullCoreName;
+      if (Platform.isWindows) {
+        if (coreName.endsWith('.dll')) {
+          fullCoreName = coreName;
+        } else if (coreName.endsWith('_libretro')) {
+          fullCoreName = '$coreName.dll';
+        } else {
+          fullCoreName = '${coreName}_libretro.dll';
+        }
+      } else if (Platform.isMacOS) {
+        if (coreName.endsWith('.dylib')) {
+          fullCoreName = coreName;
+        } else if (coreName.endsWith('_libretro')) {
+          fullCoreName = '$coreName.dylib';
+        } else {
+          fullCoreName = '${coreName}_libretro.dylib';
+        }
+      } else if (Platform.isAndroid) {
+        if (coreName.endsWith('.so')) {
+          fullCoreName = coreName;
+        } else {
+          if (coreName.contains('_android')) {
+            fullCoreName = '$coreName.so';
+          } else if (coreName.endsWith('_libretro')) {
+            fullCoreName = '${coreName}_android.so';
+          } else {
+            fullCoreName = '${coreName}_libretro_android.so';
+          }
+        }
+      } else {
+        if (coreName.endsWith('.so')) {
+          fullCoreName = coreName;
+        } else if (coreName.endsWith('_libretro')) {
+          fullCoreName = '$coreName.so';
+        } else {
+          fullCoreName = '${coreName}_libretro.so';
+        }
+      }
+
+      final corePath = path.join(coresDir, fullCoreName);
+
+      if (await File(corePath).exists()) {
+        return corePath;
+      } else {
+        return null;
+      }
+    } catch (e) {
+      _log.e('Error getting core path: $e');
+      return null;
+    }
+  }
+
+  /// Resolves the optimal directory for RetroArch cores based on the platform and installation type.
+  ///
+  /// Supports Flatpak, AppImage, and standard installation discovery on Linux.
+  static Future<String> _getRetroArchCoresDirectory(
+    EmulatorModel retroArch,
+  ) async {
+    final retroArchDir = path.dirname(retroArch.path);
+
+    if (Platform.isLinux) {
+      final homeDir = Platform.environment['HOME'] ?? '';
+
+      if (retroArch.path.contains('flatpak')) {
+        return path.join(
+          homeDir,
+          '.var/app/org.libretro.RetroArch/config/retroarch/cores',
+        );
+      }
+
+      final configCores = path.join(homeDir, '.config/retroarch/cores');
+      if (await Directory(configCores).exists()) {
+        return configCores;
+      }
+
+      return path.join(retroArchDir, 'cores');
+    } else if (Platform.isMacOS) {
+      final homeDir = ConfigService.getRealHomePath();
+      return path.join(homeDir, 'Library/Application Support/RetroArch/cores');
+    } else {
+      return path.join(retroArchDir, 'cores');
+    }
+  }
+
+  /// Checks if the default emulator (RetroArch) is currently running on the host system.
+  static Future<bool> _isDefaultEmulatorRunning() async {
+    if (Platform.isWindows) return await _isProcessRunning('retroarch.exe');
+    if (Platform.isLinux || Platform.isMacOS) {
+      return await _isProcessRunningUnix('retroarch');
+    }
+    return false;
+  }
+
+  /// Checks for a running process on Windows using the tasklist command.
+  static Future<bool> _isProcessRunning(String processName) async {
+    if (!Platform.isWindows) return false;
+
+    try {
+      final result = await Process.run('tasklist', [
+        '/FI',
+        'IMAGENAME eq $processName',
+        '/NH',
+      ]);
+      return result.stdout.toString().toLowerCase().contains(
+        processName.toLowerCase(),
+      );
+    } catch (e) {
+      _log.e('Error checking if $processName is running: $e');
+      return false;
+    }
+  }
+
+  /// Checks for a running process on Unix-like systems using the pgrep command.
+  static Future<bool> _isProcessRunningUnix(String processName) async {
+    if (!Platform.isLinux && !Platform.isMacOS) return false;
+
+    try {
+      final result = await Process.run('pgrep', ['-i', '-f', processName]);
+      return result.exitCode == 0;
+    } catch (e) {
+      _log.e('Error checking if $processName is running (unix): $e');
+      return false;
+    }
+  }
+
+  /// Handles application re-entry (foregrounding) to detect session termination.
+  static Future<void> handleAppResumed() async {
+    if (GameSessionManager.isGameLaunched) {
+      if (Platform.isLinux) return;
+
+      final isDesktop =
+          Platform.isWindows || Platform.isLinux || Platform.isMacOS;
+      final gracePeriod = isDesktop ? 10 : 2;
+      final timeSinceLaunch = DateTime.now().difference(
+        GameSessionManager.gameLaunchTime!,
+      );
+
+      if (timeSinceLaunch.inSeconds > gracePeriod) {
+        bool emulatorStillRunning = false;
+
+        if (GameSessionManager.launchedEmulatorExe != null) {
+          emulatorStillRunning = await _isProcessRunning(
+            GameSessionManager.launchedEmulatorExe!,
+          );
+        } else {
+          emulatorStillRunning = await _isDefaultEmulatorRunning();
+        }
+
+        if (!emulatorStillRunning) {
+          _log.i(
+            'GameService: Emulator process not detected after grace period. Ending session.',
+          );
+          await GameSessionManager.endGameSession();
+        }
+      }
+    } else {
+      GamepadNavigationManager.reactivate();
+    }
+  }
+
+  /// High-level emulator status check.
+  static Future<bool> isEmulatorRunning([String? processName]) async {
+    if (processName != null) return await isProcessRunning(processName);
+    return await _isDefaultEmulatorRunning();
+  }
+
+  /// Verifies if a process is running on desktop platforms.
+  static Future<bool> isProcessRunning(String processName) async {
+    if (Platform.isWindows) return await _isProcessRunning(processName);
+    if (Platform.isLinux || Platform.isMacOS) {
+      final unixName = processName.replaceAll(
+        RegExp(r'\.exe$', caseSensitive: false),
+        '',
+      );
+      return await _isProcessRunningUnix(unixName);
+    }
+    return false;
+  }
+}
